requirements.txt
Flask==2.3.3
Flask-Limiter==3.3.1
Flask-Caching==2.0.2
validators==0.20.0
aiohttp==3.8.5
asyncio==3.4.3
tldextract==3.4.4
redis==5.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-whois==0.8.0
cachetools==5.3.1